refactor: move all signUp protocols to a single file
