+++ src/presentation/controllers/index.ts
@@ -1,1 +1,1 @@
-export * from './signUp'
+export * from './signup'
